Add class_aliases
