+++ src/CoreBundle/Form/EventListener/FixCheckboxDataListener.php
@@ -13,42 +13,24 @@ declare(strict_types=1);
 
 namespace Sonata\CoreBundle\Form\EventListener;
 
-use Symfony\Component\Form\FormEvent;
-use Symfony\Component\Form\FormEvents;
+if (!class_exists(\Sonata\Form\EventListener\FixCheckboxDataListener::class, false)) {
+    @trigger_error(
+        'The '.__NAMESPACE__.'\FixCheckboxDataListener class is deprecated since version 3.13.0 and will be removed in 4.0.'
+        .' Use Sonata\Form\EventListener\FixCheckboxDataListener instead.',
+        E_USER_DEPRECATED
+    );
+}
 
-@trigger_error(
-    'The '.__NAMESPACE__.'\FixCheckboxDataListener class is deprecated since version 3.13.0 and will be removed in 4.0.'
-    .' Use Sonata\Form\EventListener\FixCheckboxDataListener instead.',
-    E_USER_DEPRECATED
+class_alias(
+    \Sonata\Form\EventListener\FixCheckboxDataListener::class,
+    __NAMESPACE__.'\FixCheckboxDataListener'
 );
 
-/**
- * @deprecated since sonata-project/core-bundle 3.13.0, to be removed in 4.0.
- */
-class FixCheckboxDataListener extends \Sonata\Form\EventListener\FixCheckboxDataListener
-{
-    public static function getSubscribedEvents()
-    {
-        return [FormEvents::PRE_SUBMIT => 'preBind'];
-    }
-
+if (false) {
     /**
-     * NEXT_MAJOR: remove this method.
-     *
-     * @deprecated since sonata-project/core-bundle 2.3, to be renamed in 4.0.
-     *             Use {@link preSubmit} instead
+     * @deprecated since sonata-project/core-bundle 3.13.0, to be removed in 4.0.
      */
-    public function preBind(FormEvent $event)
+    class FixCheckboxDataListener extends \Sonata\Form\EventListener\FixCheckboxDataListener
     {
-        // BC prevention for class extending this one.
-        if (self::class !== static::class) {
-            @trigger_error(
-                __METHOD__.' is deprecated since 2.3 and will be renamed in 4.0.'
-                .' Use '.__CLASS__.'::preSubmit instead.',
-                E_USER_DEPRECATED
-            );
-        }
-
-        $this->preSubmit($event);
     }
 }
